--- components/od-service/run.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-source "$(dirname $0)/build.sh"
-
-log_section "Starting Open Discovery Service..."
-java -jar "$COMPONENT_DIR/target/od-service-$ARTIFACT_VERSION.jar"
