+++ components/ord-service/Dockerfile
@@ -1,6 +1,6 @@
 FROM maven:3.6.3-jdk-8-slim AS builder
 
-ENV BASE_APP_DIR /go/src/github.com/kyma-incubator/compass/components/od-service
+ENV BASE_APP_DIR /go/src/github.com/kyma-incubator/compass/components/ord-service
 WORKDIR ${BASE_APP_DIR}
 
 RUN apt-get update && apt-get install -y git
@@ -17,7 +17,7 @@ COPY . .
 
 RUN ./build.sh && \
     VERSION=$(mvn org.apache.maven.plugins:maven-help-plugin:3.2.0:evaluate -Dexpression=project.version -DforceStdout -q) && \
-    mkdir /app && mv ./target/od-service-$VERSION.jar /app/od-service.jar
+    mkdir /app && mv ./target/ord-service-$VERSION.jar /app/ord-service.jar
 
 FROM openjdk:8
 LABEL source = git@github.com:kyma-incubator/compass.git
@@ -27,7 +27,7 @@ WORKDIR /app
 # Copy binary
 #
 
-COPY --from=builder /app/od-service.jar /app/
+COPY --from=builder /app/ord-service.jar /app/
 
 #
 # Install curl & CA certs
@@ -39,4 +39,4 @@ RUN apt-get install -y curl ca-certificates
 # Run app
 #
 
-CMD ["java", "-jar", "od-service.jar"]
+CMD ["java", "-jar", "ord-service.jar"]
